I moved the feilds within EventDetails Model to the Event Model. I added getters and setters for the feilds, I added the feilds to the Event constructor, and then I delete EventDetails. All of the functions to view all events, create event, delete event, and edit event are working properly.
