+++ src/main/java/TournamentOrganizer/controllers/EventController.java
@@ -2,7 +2,6 @@ package TournamentOrganizer.controllers;
 
 import TournamentOrganizer.data.EventRepository;
 import TournamentOrganizer.models.Event;
-import TournamentOrganizer.models.EventDetails;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -128,14 +127,17 @@ public class EventController {
             String competitiveLevel,
             String date,
             String summary,
-            String entryFee,
-            EventDetails eventDetails
+            String entryFee
     ) {
         Optional<Event> event = eventRepository.findById(eventId);
         Event eventToBeEdit = event.get();
         eventToBeEdit.setName(name);
-        // Name is now able to be edited, but I am still working on how to update the
-        // fields in EventDetails.
+        eventToBeEdit.setDescription(description);
+        eventToBeEdit.setLocation(location);
+        eventToBeEdit.setCompetitiveLevel(competitiveLevel);
+        eventToBeEdit.setDate(date);
+        eventToBeEdit.setSummary(summary);
+        eventToBeEdit.setEntryFee(entryFee);
         eventRepository.save(eventToBeEdit);
         String redirect = "redirect:";
         return redirect;
