--- src/main/resources/templates/events/detail.html
@@ -1,44 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org/">
-<head th:replace="fragments :: head"></head>
-<body class="container">
-<div th:replace="fragments :: page-header"></div>
-
-<h1 th:replace="fragments :: title" class="text-warning"></h1>
-
-<nav th:replace="fragments :: navigation"></nav>
-
-<table class="table table-bordered" style="background-color: maroon">
-    <tr>
-        <th>ID</th>
-        <td th:text="${event.id}"></td>
-    </tr>
-    <tr>
-        <th>Description</th>
-        <td th:text="${event.eventDetails.description}"></td>
-    </tr>
-    <tr>
-        <th>Location</th>
-        <td th:text="${event.eventDetails.location}"></td>
-    </tr>
-    <tr>
-        <th>Competitive Level</th>
-        <td th:text="${event.eventDetails.competitiveLevel}"></td>
-    </tr>
-    <tr>
-        <th>Date</th>
-        <td th:text="${event.eventDetails.date}"></td>
-    </tr>
-    <tr>
-        <th>Summary</th>
-        <td th:text="${event.eventDetails.summary}"></td>
-    </tr>
-    <tr>
-        <th>Entry Fee</th>
-        <td th:text="${event.eventDetails.entryFee}"></td>
-    </tr>
-
-</table>
-
-</body>
-</html>
